Move displayit() to clear_display.c
Furthermore make 'terminal1..4' strings local
